0209_승현_광고기능

+++ pawject4/pawject4_back/src/main/java/com/pawject/dto/tester/TesterAdminResponseDto.java
@@ -45,9 +45,7 @@ public class TesterAdminResponseDto {
                 .toList();
 
         return TesterAdminResponseDto.builder()
-        		
-        	    .testerid(t.getTesterid() == null ? null : t.getTesterid().longValue())
-
+        	    .testerid(t.getTesterid() == null ? null : t.getTesterid())
         	    .category(t.getCategory())
         	    .title(t.getTitle())
         	    .content(t.getContent())
